Clean up misc
