+++ src/web/package.json
@@ -66,6 +66,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-modal": "^3.15.1",
+    "reselect": "^4.1.6",
     "saxon-js": "^2.3.0",
     "uswds": "^2.12.0",
     "xml-formatter": "^2.6.1"
