+++ src/web/src/browser/presenter/state/validation-results-machine.ts
@@ -3,8 +3,6 @@ import type {
   ValidationReport,
 } from '@asap/shared/use-cases/schematron';
 
-import { getAssertionsById } from '../lib/validator';
-
 type BaseState = {
   summary: {
     firedCount: null | number;
@@ -55,6 +53,22 @@ export type StateTransition =
   | {
       type: 'RESET';
     };
+
+export type FailedAssertionMap = Record<FailedAssert['id'], FailedAssert[]>;
+
+export const getAssertionsById = ({
+  failedAssertions,
+}: {
+  failedAssertions: FailedAssert[];
+}) => {
+  return failedAssertions.reduce((acc, assert) => {
+    if (acc[assert.id] === undefined) {
+      acc[assert.id] = [];
+    }
+    acc[assert.id].push(assert);
+    return acc;
+  }, {} as FailedAssertionMap);
+};
 
 export const nextState = (state: State, event: StateTransition): State => {
   if (state.current === 'NO_RESULTS') {
